streamlining the database cleaning procedure

+++ python/clean_database.py
@@ -2,6 +2,7 @@ from cdp_backend.database import models as db_models
 import fireo
 from google.auth.credentials import AnonymousCredentials
 from google.cloud.firestore import Client
+import sys
 
 # Connect to the database
 fireo.connection(client=Client(
@@ -12,14 +13,6 @@ fireo.connection(client=Client(
 # Read from the database
 sessions = db_models.Session.collection.fetch(1000)
 
-i=0
-for session in sessions:
-    if session.session_datetime.year == 2020 and session.session_datetime.month==1 and session.session_datetime.day==7:
-        print("Session_time=" + str(session.session_datetime) + " ID=" + str(session.id) + " key=" +str(session.key) +" _update_time=" + str(session._update_time))
-    i=i+1
-
-print("----")
-
 dictionary = dict()
 sessions_for_delete = []
 sessions = db_models.Session.collection.fetch(1000)
@@ -27,7 +20,6 @@ for session in sessions:
     if session.session_datetime not in dictionary:
         dictionary[session.session_datetime] = []
     dictionary[session.session_datetime].append(session)
-i = 0
 
 # Finding the sessions to delete
 for key in sorted(dictionary.keys()):
@@ -36,13 +28,16 @@ for key in sorted(dictionary.keys()):
         for session in dictionary[key]:
             if session.session_datetime.year == 2020 and session._update_time.year == 2022:
                 sessions_for_delete.append(session)
-                print("Session_time=" + str(session.session_datetime) + " ID=" + str(session.id) + " key=" + str(session.key) + " _update_time=" + str(session._update_time) + " delete this from 2020")
+                print(f"Session_time={str(session.session_datetime)} ID={str(session.id)} key={str(session.key)} update_time={str(session._update_time)} delete this from 2020")
             else:
-                print("Session_time=" + str(session.session_datetime) + " ID=" + str(session.id) + " key=" + str(session.key) + " _update_time=" + str(session._update_time))
+                print(f"Session_time={str(session.session_datetime)} ID={str(session.id)} key={str(session.key)} update_time={str(session._update_time)}")
         print("")
-        i = i + 1
-print(i)
 
 # This will delete the actual sessions
-for session in sessions_for_delete:
-    db_models.Session.collection.delete(session.key)
+if sys.argv[-1] == "real":
+    for session in sessions_for_delete:
+        db_models.Session.collection.delete(session.key)
+elif sys.argv[-1] == "dry_run":
+    print(f"This is a dry run, but if it were real, {len(sessions_for_delete)} duplicates would be deleted.")
+else:
+    print("Please pass a valid parameter: \"real\" or \"dry_run.\"")
